Calculator (1/28/17)
Initial creation of the program. Currently has 4 functions for each corresponding arithmetic equation. Also has a random number generator(int) that takes a range. The "main" function is the main menu and "Choices" function is where the choices are evaluated. Used a global variable for deciding which number for the choice was chosen. 1/28/17